answer_machine/requirements.txt
numpy==1.24.4
torch==1.13.1
transformers==4.30.2
accelerate==0.18.0
tqdm
scikit-learn
psycopg2-binary
